X Separation of concerns for buffer*
...mirror, view and server side buffer.

* buffer.version changed from i32 to usize
* Removed the buffer_mirror.selections
  HashMap<BufferViewID, Selection> and made it a singular
  Selection. Not a single reference to the field leveraged
  the collection, whist adding a lot of complexity and confusion
  in how it should be handled. view_id doesn't need to be
  sent as an argument to neither set_selection() nor
  selection(), for example. Also led to a wierd API where
  contributors where getting buffer_view.view_ids by
  getting all the keys of buffer_mirror.selections.

  Deprecates selections() method, removed it along with reset_selections.
  It also makes it clear that Selection is always set, deprecating
  ensure_view_init() and remove_view() in the process.

  Most commands retrieved buffer
  just to call:

  let (buffer_view, buffer_mirror) = current!(ui_tree)
  selection = buffer_mirror.selection(buffer_view.view_id)

  All in all, the changes could lead to the same thing being accomplished by:

  let buffer_mirror = current!(ui_tree)
  selection = buffer_mirror.selection()

  current_ref! and current! have therefore also been modified to only return
  buffer_mirror, deprecating doc! and doc_mut!.

  (sure there are some occasions where the view is also updated, but there are
   some fields that will be moved from view to buffer_mirror. The occurence of
   such occasions are in other words planned to be reduced.)

  (buffer_mirror most likely to be changed to somehing less verbose, given its
   frequency.)

* Renamed current_ref to current and current to current_mut

+++ helix-lsp/src/client.rs
@@ -430,7 +430,7 @@ impl Client {
     pub fn text_document_did_open(
         &self,
         uri: lsp::Url,
-        version: i32,
+        version: usize,
         doc: &Rope,
         language_id: String,
     ) -> impl Future<Output = Result<()>> {
